requirements.txt
keras
IPython
Pillow
